Create new file to hold Settings logic, fix unsafe memory usage
